Slim I3CompressedPhoton further.
Current size is 56 bytes, including 4 bytes of padding required to align
the double and int64 fields. Storing the time in single precision (which
is potentially dangerous as noted in #1978), would reduce it to 48 bytes,
and omitting the source particle IDs would bring it down to an absolute
minimum of 36 bytes. (See #1978)


git-svn-id: http://code.icecube.wisc.edu/svn/projects/simclasses/trunk@154595 16731396-06f5-0310-8873-f7f720988828